fix: polish pay modal
- move risk checkbox content to own component
- clean up spacing for mint ERC-20 checkbox
- fix link to docs section in risks
